Git status tests moved to a separate test module.
